+++ source/GameData.h
@@ -32,6 +32,7 @@ class Effect;
 class Fleet;
 class Galaxy;
 class GameEvent;
+class GameObjects;
 class Government;
 class Hazard;
 class ImageSet;
@@ -63,18 +64,19 @@ class TestData;
 // universe.
 class GameData {
 public:
-	static bool BeginLoad(const char * const *argv);
+	static void BeginLoad(bool onlyLoadData, bool debugMode);
+	static void FinishLoading();
 	// Check for objects that are referred to but never defined.
 	static void CheckReferences();
 	static void LoadShaders(bool useShaderSwizzle);
-	// TODO: make Progress() a simple accessor.
 	static double Progress();
-	// Whether initial game loading is complete (sprites and audio are loaded).
+	// Whether initial game loading is complete (data, sprites and audio are loaded).
 	static bool IsLoaded();
+	static bool IsDataLoaded();
 	// Begin loading a sprite that was previously deferred. Currently this is
 	// done with all landscapes to speed up the program's startup.
 	static void Preload(const Sprite *sprite);
-	static void FinishLoading();
+	static void FinishLoadingSprites();
 	
 	// Get the list of resource sources (i.e. plugin folders).
 	static const std::vector<std::string> &Sources();
@@ -156,12 +158,7 @@ public:
 	
 private:
 	static void LoadSources();
-	static void LoadFile(const std::string &path, bool debugMode);
 	static std::map<std::string, std::shared_ptr<ImageSet>> FindImages();
-	
-	static void PrintShipTable();
-	static void PrintTestsTable();
-	static void PrintWeaponTable();
 };
