+++ source/StarField.h
@@ -35,6 +35,7 @@ class StarField {
 public:
 	void Init(int stars, int width);
 	void SetHaze(const Sprite *sprite, bool allowAnimation);
+	void ResetAnimation();
 	
 	void Draw(const Point &pos, const Point &vel, double zoom = 1.) const;
