+++ source/main.cpp
@@ -23,6 +23,7 @@ PARTICULAR PURPOSE.  See the GNU General Public License for more details.
 #include "text/Font.h"
 #include "FrameTimer.h"
 #include "GameData.h"
+#include "GameObjects.h"
 #include "GameWindow.h"
 #include "MenuPanel.h"
 #include "Panel.h"
@@ -37,6 +38,7 @@ PARTICULAR PURPOSE.  See the GNU General Public License for more details.
 #include <chrono>
 #include <iostream>
 #include <map>
+#include <thread>
 
 #include <stdexcept>
 #include <string>
@@ -51,7 +53,10 @@ PARTICULAR PURPOSE.  See the GNU General Public License for more details.
 using namespace std;
 
 void PrintHelp();
+void PrintShipTable();
+void PrintTestsTable();
 void PrintVersion();
+void PrintWeaponTable();
 void GameLoop(PlayerInfo &player, const Conversation &conversation, const string &testToRun, bool debugMode);
 Conversation LoadConversation();
 #ifdef _WIN32
@@ -71,8 +76,12 @@ int main(int argc, char *argv[])
 	Conversation conversation;
 	bool debugMode = false;
 	bool loadOnly = false;
+	bool printShips = false;
+	bool printTests = false;
+	bool printWeapons = false;
 	string testToRunName = "";
 
+	Files::Init(argv);
 	for(const char *const *it = argv + 1; *it; ++it)
 	{
 		string arg = *it;
@@ -86,6 +95,8 @@ int main(int argc, char *argv[])
 			PrintVersion();
 			return 0;
 		}
+		else if(arg == "-s" || arg == "--ships")
+			printShips = true;
 		else if(arg == "-t" || arg == "--talk")
 			conversation = LoadConversation();
 		else if(arg == "-d" || arg == "--debug")
@@ -94,28 +105,53 @@ int main(int argc, char *argv[])
 			loadOnly = true;
 		else if(arg == "--test" && *++it)
 			testToRunName = *it;
+		else if(arg == "--tests")
+			printTests = true;
+		else if(arg == "-w" || arg == "--weapons")
+			printWeapons = true;
 	}
 	
 	try {
-		// Begin loading the game data. Exit early if we are not using the UI.
-		if(!GameData::BeginLoad(argv))
+		// Begin loading the game data.
+		GameData::BeginLoad(loadOnly || printShips || printTests || printWeapons, debugMode);
+
+		PlayerInfo player;
+
+		// Exit early if we are not using the UI.
+		if(printShips || printTests || printWeapons)
+		{
+			while(!GameData::IsDataLoaded())
+				this_thread::yield();
+
+			if(printShips)
+				PrintShipTable();
+			if(printTests)
+				PrintTestsTable();
+			if(printWeapons)
+				PrintWeaponTable();
 			return 0;
+		}
+		if(loadOnly)
+		{
+			while(!GameData::IsDataLoaded())
+				this_thread::yield();
+			GameData::FinishLoading();
+
+			// Load player data, including reference-checking.
+			bool checkedReferences = player.LoadRecent();
+			if(loadOnly)
+			{
+				if(!checkedReferences)
+					GameData::CheckReferences();
+				cout << "Parse completed." << endl;
+			}
+			return 0;
+		}
 		
 		if(!testToRunName.empty() && !GameData::Tests().Has(testToRunName))
 		{
 			Files::LogError("Test \"" + testToRunName + "\" not found.");
 			return 1;
-		}
-		
-		// Load player data, including reference-checking.
-		PlayerInfo player;
-		bool checkedReferences = player.LoadRecent();
-		if(loadOnly)
-		{
-			if(!checkedReferences)
-				GameData::CheckReferences();
-			cout << "Parse completed." << endl;
-			return 0;
 		}
 		
 		// On Windows, make sure that the sleep timer has at least 1 ms resolution
@@ -133,8 +169,6 @@ int main(int argc, char *argv[])
 		
 		// Show something other than a blank window.
 		GameWindow::Step();
-		
-		Audio::Init(GameData::Sources());
 		
 		// This is the main loop where all the action begins.
 		GameLoop(player, conversation, testToRunName, debugMode);
@@ -359,6 +393,109 @@ void PrintHelp()
 
 
 
+// This prints out the list of tests that are available and their status
+// (active/missing feature/known failure)..
+void PrintTestsTable()
+{
+	cout << "status" << '\t' << "name" << '\n';
+	for(auto &it : GameData::Tests())
+	{
+		const Test &test = it.second;
+		cout << test.StatusText() << '\t';
+		cout << "\"" << test.Name() << "\"" << '\n';
+	}
+	cout.flush();
+}
+
+
+
+void PrintShipTable()
+{
+	cout << "model" << '\t' << "cost" << '\t' << "shields" << '\t' << "hull" << '\t'
+		<< "mass" << '\t' << "crew" << '\t' << "cargo" << '\t' << "bunks" << '\t'
+		<< "fuel" << '\t' << "outfit" << '\t' << "weapon" << '\t' << "engine" << '\t'
+		<< "speed" << '\t' << "accel" << '\t' << "turn" << '\t'
+		<< "energy generation" << '\t' << "max energy usage" << '\t' << "energy capacity" << '\t'
+		<< "idle/max heat" << '\t' << "max heat generation" << '\t' << "max heat dissipation" << '\t'
+		<< "gun mounts" << '\t' << "turret mounts" << '\n';
+	for(auto &it : GameData::Ships())
+	{
+		// Skip variants and unnamed / partially-defined ships.
+		if(it.second.ModelName() != it.first)
+			continue;
+
+		const Ship &ship = it.second;
+		cout << it.first << '\t';
+		cout << ship.Cost() << '\t';
+
+		const Outfit &attributes = ship.Attributes();
+		auto mass = attributes.Mass() ? attributes.Mass() : 1.;
+		cout << attributes.Get("shields") << '\t';
+		cout << attributes.Get("hull") << '\t';
+		cout << mass << '\t';
+		cout << attributes.Get("required crew") << '\t';
+		cout << attributes.Get("cargo space") << '\t';
+		cout << attributes.Get("bunks") << '\t';
+		cout << attributes.Get("fuel capacity") << '\t';
+
+		cout << ship.BaseAttributes().Get("outfit space") << '\t';
+		cout << ship.BaseAttributes().Get("weapon capacity") << '\t';
+		cout << ship.BaseAttributes().Get("engine capacity") << '\t';
+		cout << (attributes.Get("drag") ? (60. * attributes.Get("thrust") / attributes.Get("drag")) : 0) << '\t';
+		cout << 3600. * attributes.Get("thrust") / mass << '\t';
+		cout << 60. * attributes.Get("turn") / mass << '\t';
+
+		double energyConsumed = attributes.Get("energy consumption")
+			+ max(attributes.Get("thrusting energy"), attributes.Get("reverse thrusting energy"))
+			+ attributes.Get("turning energy")
+			+ attributes.Get("afterburner energy")
+			+ attributes.Get("fuel energy")
+			+ (attributes.Get("hull energy") * (1 + attributes.Get("hull energy multiplier")))
+			+ (attributes.Get("shield energy") * (1 + attributes.Get("shield energy multiplier")))
+			+ attributes.Get("cooling energy")
+			+ attributes.Get("cloaking energy");
+
+		double heatProduced = attributes.Get("heat generation") - attributes.Get("cooling")
+			+ max(attributes.Get("thrusting heat"), attributes.Get("reverse thrusting heat"))
+			+ attributes.Get("turning heat")
+			+ attributes.Get("afterburner heat")
+			+ attributes.Get("fuel heat")
+			+ (attributes.Get("hull heat") * (1 + attributes.Get("hull heat multiplier")))
+			+ (attributes.Get("shield heat") * (1 + attributes.Get("shield heat multiplier")))
+			+ attributes.Get("solar heat")
+			+ attributes.Get("cloaking heat");
+
+		for(const auto &oit : ship.Outfits())
+			if(oit.first->IsWeapon() && oit.first->Reload())
+			{
+				double reload = oit.first->Reload();
+				energyConsumed += oit.second * oit.first->FiringEnergy() / reload;
+				heatProduced += oit.second * oit.first->FiringHeat() / reload;
+			}
+		cout << 60. * (attributes.Get("energy generation") + attributes.Get("solar collection")) << '\t';
+		cout << 60. * energyConsumed << '\t';
+		cout << attributes.Get("energy capacity") << '\t';
+		cout << ship.IdleHeat() / max(1., ship.MaximumHeat()) << '\t';
+		cout << 60. * heatProduced << '\t';
+		// Maximum heat is 100 degrees per ton. Bleed off rate is 1/1000 per 60th of a second, so:
+		cout << 60. * ship.HeatDissipation() * ship.MaximumHeat() << '\t';
+
+		int numTurrets = 0;
+		int numGuns = 0;
+		for(auto &hardpoint : ship.Weapons())
+		{
+			if(hardpoint.IsTurret())
+				++numTurrets;
+			else
+				++numGuns;
+		}
+		cout << numGuns << '\t' << numTurrets << '\n';
+	}
+	cout.flush();
+}
+
+
+
 void PrintVersion()
 {
 	cerr << endl;
@@ -369,6 +506,48 @@ void PrintVersion()
 	cerr << endl;
 	cerr << GameWindow::SDLVersions() << endl;
 	cerr << endl;
+}
+
+
+
+void PrintWeaponTable()
+{
+	cout << "name" << '\t' << "cost" << '\t' << "space" << '\t' << "range" << '\t'
+		<< "energy/s" << '\t' << "heat/s" << '\t' << "recoil/s" << '\t'
+		<< "shield/s" << '\t' << "hull/s" << '\t' << "push/s" << '\t'
+		<< "homing" << '\t' << "strength" <<'\n';
+	for(auto &it : GameData::Outfits())
+	{
+		// Skip non-weapons and submunitions.
+		if(!it.second.IsWeapon() || it.second.Category().empty())
+			continue;
+
+		const Outfit &outfit = it.second;
+		cout << it.first << '\t';
+		cout << outfit.Cost() << '\t';
+		cout << -outfit.Get("weapon capacity") << '\t';
+
+		cout << outfit.Range() << '\t';
+
+		double energy = outfit.FiringEnergy() * 60. / outfit.Reload();
+		cout << energy << '\t';
+		double heat = outfit.FiringHeat() * 60. / outfit.Reload();
+		cout << heat << '\t';
+		double firingforce = outfit.FiringForce() * 60. / outfit.Reload();
+		cout << firingforce << '\t';
+
+		double shield = outfit.ShieldDamage() * 60. / outfit.Reload();
+		cout << shield << '\t';
+		double hull = outfit.HullDamage() * 60. / outfit.Reload();
+		cout << hull << '\t';
+		double hitforce = outfit.HitForce() * 60. / outfit.Reload();
+		cout << hitforce << '\t';
+
+		cout << outfit.Homing() << '\t';
+		double strength = outfit.MissileStrength() + outfit.AntiMissile();
+		cout << strength << '\n';
+	}
+	cout.flush();
 }
